Only replace puppet in puppet.rb if it's present

+++ puppet/modules/jenkins_job_builder/files/theforeman.org/scripts/test/foreman-proxy.sh
@@ -18,7 +18,9 @@ set -x
 gem install bundler -v '< 2.0' --no-ri --no-rdoc
 
 # Puppet environment
-sed -i "/^\s*gem.*puppet/ s/\$/, '~> $puppet'/" $APP_ROOT/bundler.d/puppet.rb
+if [ -e $APP_ROOT/bundler.d/puppet.rb ] ; then
+	sed -i "/^\s*gem.*puppet/ s/\$/, '~> $puppet'/" $APP_ROOT/bundler.d/puppet.rb
+fi
 
 bundle install --without=development --jobs=5 --retry=5
 bundle exec rake pkg:generate_source jenkins:unit --trace
